style: Moving the ForeignKeyAssociationException to the entities Module
