+++ src/main/java/br/devrafaelsoares/storeapirestful/services/CategoryService.java
@@ -3,7 +3,7 @@ package br.devrafaelsoares.storeapirestful.services;
 import br.devrafaelsoares.storeapirestful.domain.category.Category;
 import br.devrafaelsoares.storeapirestful.domain.category.dto.CategoryCreateRequest;
 import br.devrafaelsoares.storeapirestful.domain.category.dto.CategoryUpdate;
-import br.devrafaelsoares.storeapirestful.exceptions.ForeignKeyAssociationException;
+import br.devrafaelsoares.storeapirestful.exceptions.entity.ForeignKeyAssociationException;
 import br.devrafaelsoares.storeapirestful.repositories.CategoryRepository;
 import jakarta.persistence.EntityExistsException;
 import jakarta.persistence.EntityNotFoundException;
